Cleaned up redundant code left from the business logic refactoring
