+++ workshops/introduction_to_event_sourcing/src/05_business_logic/immutable/core.ts
