+++ workshops/introduction_to_event_sourcing/src/solved/05_business_logic/immutable/solution2/core.ts
