+++ workshops/introduction_to_event_sourcing/src/solved/05_business_logic/oop/solution1/core.ts
