--- workshops/introduction_to_event_sourcing/src/solved/05_command_definition copy/oop/solution1/businessLogic.ts
@@ -1,116 +1,0 @@
-import { Aggregate, Event, EventStore } from './core';
-import { ShoppingCart, PricedProductItem } from './shoppingCart';
-
-export interface Repository<Entity> {
-  find(id: string): Entity;
-  store(id: string, entity: Entity): void;
-}
-
-export class EventStoreRepository<
-  Entity extends Aggregate<StreamEvent>,
-  StreamEvent extends Event,
-> implements Repository<Entity>
-{
-  constructor(
-    private eventStore: EventStore,
-    private getInitialState: () => Entity,
-  ) {}
-
-  find = (id: string): Entity => {
-    const currentState = this.getInitialState();
-
-    const events = this.eventStore.readStream<StreamEvent>(id);
-
-    for (const event of events) {
-      currentState.evolve(event);
-    }
-
-    return currentState;
-  };
-
-  store = (id: string, entity: Entity): void => {
-    const events = entity.dequeueUncommitedEvents();
-
-    if (events.length === 0) return;
-
-    this.eventStore.appendToStream(id, ...events);
-  };
-}
-
-export abstract class ApplicationService<Entity> {
-  constructor(protected repository: Repository<Entity>) {}
-
-  protected on = (id: string, handle: (state: Entity) => void) => {
-    const aggregate = this.repository.find(id);
-
-    handle(aggregate);
-
-    this.repository.store(id, aggregate);
-  };
-}
-
-export type OpenShoppingCart = {
-  shoppingCartId: string;
-  clientId: string;
-  now: Date;
-};
-
-export type AddProductItemToShoppingCart = {
-  shoppingCartId: string;
-  productItem: PricedProductItem;
-};
-
-export type RemoveProductItemFromShoppingCart = {
-  shoppingCartId: string;
-  productItem: PricedProductItem;
-};
-
-export type ConfirmShoppingCart = {
-  shoppingCartId: string;
-  now: Date;
-};
-
-export type CancelShoppingCart = {
-  shoppingCartId: string;
-  now: Date;
-};
-
-export type ShoppingCartCommand =
-  | OpenShoppingCart
-  | AddProductItemToShoppingCart
-  | RemoveProductItemFromShoppingCart
-  | ConfirmShoppingCart
-  | CancelShoppingCart;
-
-export class ShoppingCartService extends ApplicationService<ShoppingCart> {
-  constructor(protected repository: Repository<ShoppingCart>) {
-    super(repository);
-  }
-
-  public open = ({ shoppingCartId, clientId, now }: OpenShoppingCart) =>
-    this.on(shoppingCartId, (shoppingCart) =>
-      shoppingCart.open(shoppingCartId, clientId, now),
-    );
-
-  public addProductItem = ({
-    shoppingCartId,
-    productItem,
-  }: AddProductItemToShoppingCart) =>
-    this.on(shoppingCartId, (shoppingCart) =>
-      shoppingCart.addProductItem(productItem),
-    );
-
-  public removeProductItem = ({
-    shoppingCartId,
-    productItem,
-  }: RemoveProductItemFromShoppingCart) =>
-    this.on(shoppingCartId, (shoppingCart) =>
-      shoppingCart.removeProductItem(productItem),
-    );
-
-  public confirm = ({ shoppingCartId, now }: ConfirmShoppingCart) =>
-    this.on(shoppingCartId, (shoppingCart) => shoppingCart.confirm(now));
-
-  public cancel = ({ shoppingCartId, now }: CancelShoppingCart) =>
-    this.on(shoppingCartId, (shoppingCart) => shoppingCart.cancel(now));
-}
