--- workshops/introduction_to_event_sourcing/src/solved/05_command_definition/oop/solution1/core.ts
@@ -1,45 +1,0 @@
-export type Event<
-  EventType extends string = string,
-  EventData extends Record<string, unknown> = Record<string, unknown>,
-> = Readonly<{
-  type: Readonly<EventType>;
-  data: Readonly<EventData>;
-}>;
-
-export abstract class Aggregate<E extends Event> {
-  #uncommitedEvents: E[] = [];
-
-  abstract evolve(event: E): void;
-
-  protected enqueue = (event: E) => {
-    this.#uncommitedEvents = [...this.#uncommitedEvents, event];
-  };
-
-  dequeueUncommitedEvents = (): Event[] => {
-    const events = this.#uncommitedEvents;
-
-    this.#uncommitedEvents = [];
-
-    return events;
-  };
-}
-
-export interface EventStore {
-  readStream<E extends Event>(streamId: string): E[];
-  appendToStream(streamId: string, ...events: Event[]): void;
-}
-
-export const getEventStore = () => {
-  const streams = new Map<string, Event[]>();
-
-  return {
-    readStream: <E extends Event>(streamId: string): E[] => {
-      return streams.get(streamId)?.map((e) => <E>e) ?? [];
-    },
-    appendToStream: (streamId: string, ...events: Event[]): void => {
-      const current = streams.get(streamId) ?? [];
-
-      streams.set(streamId, [...current, ...events]);
-    },
-  };
-};
